Update Command Bus and move Bus to application layer
