+++ src/SalaryHistory/Application/UseCases/Commands/StoreSalaryHistoryCommand.php
@@ -3,8 +3,9 @@
 namespace Src\SalaryHistory\Application\UseCases\Commands;
 
 use Src\SalaryHistory\Application\DTOs\StoreSalaryHistoryDTO;
+use Src\Shared\Application\ICommand;
 
-class StoreSalaryHistoryCommand
+class StoreSalaryHistoryCommand implements ICommand
 {
     public function __construct(
         public StoreSalaryHistoryDTO $dto
