+++ src/SalaryHistory/Application/UseCases/Commands/UpdateSalaryHistoryCommand.php
@@ -3,8 +3,9 @@
 namespace Src\SalaryHistory\Application\UseCases\Commands;
 
 use Src\SalaryHistory\Application\DTOs\UpdateSalaryHistoryDTO;
+use Src\Shared\Application\ICommand;
 
-class UpdateSalaryHistoryCommand
+class UpdateSalaryHistoryCommand implements ICommand
 {
     public function __construct(
         public UpdateSalaryHistoryDTO $dto
