+++ src/SalaryHistory/Presentation/SalaryHistoryController.php
@@ -15,7 +15,6 @@ use Src\SalaryHistory\Presentation\Requests\StoreSalaryHistoryRequest;
 use Src\SalaryHistory\Application\UseCases\Commands\StoreSalaryHistoryCommand;
 use Src\SalaryHistory\Application\UseCases\Commands\UpdateSalaryHistoryCommand;
 use Src\SalaryHistory\Application\UseCases\Queries\GetSalaryHistoriesQuery;
-use Src\SalaryHistory\Application\UseCases\Queries\InvalidQuery;
 use Src\SalaryHistory\Infrastructure\Buses\CommandBus;
 use Src\SalaryHistory\Presentation\Requests\GetSalaryHistoryRequest;
 use Src\SalaryHistory\Presentation\Requests\UpdateSalaryHistoryRequest;
