+++ src/Shared/Application/ICommandHandler.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Src\Shared\Application;
+
+interface ICommandHandler
+{
+    /**
+     * @param ICommand $command
+     * 
+     * @return mixed
+     */
+    public function handle(ICommand $command);
+}
