+++ src/Shared/Application/IQuery.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Src\Shared\Application;
+
+interface IQuery
+{
+}
