--- src/Shared/Domain/IQuery.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Src\Shared\Domain;
-
-interface IQuery
-{
-}
